requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
python-magic==0.4.15
requests==2.24.0
flask==1.1.1
flask-uploads==0.2.1
Werkzeug==0.16.0
numpy==1.16.6 #1.17.2
Pillow==7.0.0
matplotlib==3.1.3
#opencv-contrib-python==3.4.2.17
torch==1.6.0+cpu
torchvision==0.7.0+cpu
scipy==1.2.1
scikit-learn==0.21.0